Corresponding to the material separated model data.
The shader uses only 2 IBL texture(irradiance and specular), and not
reference the metallic and roughness in the texture.
The irradiance is used the same texture at any material.
The specular texture is decided by the roughness of each material.
It decreases the flexibility but increases the efficiency of the shader.