Optimise a sea shader.
dot(col, vec3(-2.0, 1.0, 1.0) is a trick for coast.ktx.
Each terrain element has different color.
I want to apply the noise to the water and woods.
They have weak red component, and the others have strong one.
Therefore, we can control the area we want to apply the noise by the red component.