Implement ATiTC decorder.
The textures in Others directory and in Adreno directory is almost same.
therefore, it is sufficient only one side.
I focus Adreno GPU(because i love my 003SH), so I want to keep Adreno side,
but other GPU can not read ATITC format.
For that reason, I implement this  function.